LayoutTests/editing/smart-lists/newline-on-empty-list-element-removes-plain-text-markers-expected.txt

Pressing Enter on an empty list item (double newline) should break out of the list, moving the following text "C" outside the list as its own block.
| <ul>
|   class="Apple-disc-list"
|   style="list-style-type: disc;"
|   <li>
|     "A"
|     <br>
|   <li>
|     "B"
| <div>
|   "C<#selection-caret>"
